refactor: remove cm5 graphql
