+++ libs/frontend/view/components/codeMirror/CodeMirrorEditor.tsx
@@ -1,19 +1,17 @@
-import 'eslint-linter-browserify'
 import { CodeMirrorLanguage } from '@codelab/shared/abstract/codegen'
 import { css } from '@codemirror/lang-css'
 import { javascript } from '@codemirror/lang-javascript'
 import { json } from '@codemirror/lang-json'
 import { StreamLanguage } from '@codemirror/language'
-import { Extension } from '@uiw/react-codemirror'
 import { graphql } from 'codemirror-graphql/cm6-legacy/mode'
 import React from 'react'
-import { CodeMirrorInput, CodeMirrorInputProps } from './codeMirrorInput'
+import { CodeMirrorInput, CodeMirrorInputProps } from './CodeMirrorInput'
 
 export interface CodeMirrorEditorProps extends CodeMirrorInputProps {
   language?: CodeMirrorLanguage
 }
 
-const languageExtension: { [lang in CodeMirrorLanguage]: Array<Extension> } = {
+const languageExtension = {
   [CodeMirrorLanguage.Css]: [css()],
   [CodeMirrorLanguage.Javascript]: [javascript()],
   [CodeMirrorLanguage.Typescript]: [javascript({ typescript: true })],
@@ -30,7 +28,7 @@ export const CodeMirrorEditor = (props: CodeMirrorEditorProps) => {
       // eslint-disable-next-line react/jsx-props-no-spreading
       {...props}
       extensions={
-        language ? languageExtension[language].concat(extensions) : extensions
+        language ? [...languageExtension[language], ...extensions] : extensions
       }
       shouldDisableNewLines={false}
     />
