+++ libs/frontend/view/components/codeMirror/index.ts
@@ -1,0 +1,2 @@
+export * from './CodeMirrorEditor'
+export * from './CodeMirrorInput'
