--- libs/frontend/view/components/editors/codeMirrorInput/index.ts
@@ -1,2 +1,0 @@
-export * from './CodeMirrorInput'
-export * from './extensions'
