--- libs/frontend/view/components/editors/graphqlEditor/ReactCodeMirror/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ReactCodeMirror'
