--- libs/frontend/view/components/editors/index.ts
@@ -1,3 +1,0 @@
-export * from './CodeMirrorEditor'
-export * from './codeMirrorInput'
-export * from './graphqlEditor'
