+++ package.json
@@ -50,11 +50,11 @@
     "@aws-sdk/client-lambda": "3.51.0",
     "@aws-sdk/client-s3": "3.51.0",
     "@babel/runtime": "7.18.3",
-    "@codemirror/autocomplete": "0.20.0",
-    "@codemirror/lang-css": "^0.20.0",
-    "@codemirror/lang-html": "^0.20.0",
-    "@codemirror/lang-javascript": "^0.20.0",
-    "@codemirror/lang-json": "^0.20.0",
+    "@codemirror/autocomplete": "6.1.0",
+    "@codemirror/lang-css": "^6.0.0",
+    "@codemirror/lang-html": "^6.1.0",
+    "@codemirror/lang-javascript": "^6.0.2",
+    "@codemirror/lang-json": "^6.0.0",
     "@codemirror/legacy-modes": "^6.1.0",
     "@dnd-kit/core": "^6.0.5",
     "@emotion/react": "11.10.0",
@@ -89,8 +89,7 @@
     "@react-hook/window-size": "^3.0.7",
     "@slack/web-api": "6.5.1",
     "@storybook/core-server": "6.5.9",
-    "@uiw/react-codemirror": "4.7.0",
-    "@uiw/react-codemirror3": "npm:@uiw/react-codemirror@3.2.8",
+    "@uiw/react-codemirror": "4.11.6",
     "adm-zip": "0.5.9",
     "ajv": "8.8.2",
     "animate.css": "4.1.1",
@@ -101,7 +100,6 @@
     "body-parser": "1.19.1",
     "class-transformer": "0.5.1",
     "class-validator": "0.13.2",
-    "codemirror": "5",
     "codemirror-graphql": "^1.3.2",
     "core-js": "3.20.2",
     "cors": "2.8.5",
